+++ joulescope_ui/command.py
@@ -18,8 +18,8 @@ and saves the raw stream data to a file.  This executable is a
 development tool and is not intended for customer use.
 """
 
-import logging
 from joulescope_ui.main import run
+from joulescope_ui.logging_util import LEVELS
 
 
 NAME = "ui"
@@ -29,8 +29,11 @@ def parser_config(p):
     """Start the Joulescope graphical user interface"""
     p.add_argument('--device_name',
                    help='The device name to search [joulescope]')
+    p.add_argument('--log_level',
+                   choices=list(LEVELS.keys()),
+                   help='The console (stdout) log level.')
     return on_cmd
 
 
 def on_cmd(args):
-    return run(args.device_name)
+    return run(device_name=args.device_name, log_level=args.log_level)