+++ joulescope_ui/main.py
@@ -926,10 +926,19 @@ def kick_bootloaders():
         d.go()
 
 
-def run(device_name=None):
+def run(device_name=None, log_level=None):
+    """Run the Joulescope UI application.
+
+    :param device_name: The optional Joulescope device name.  None (default)
+        searches for normal Joulescope devices.
+    :param log_level: The logging level for the stdout console stream log.
+        The allowed levels are in :data:`joulescope_ui.logging_util.LEVELS`.
+    :return: 0 on success or error code on failure.
+    """
     cfg_def = load_config_def()
     cfg = load_config(cfg_def)
-    logging_config(file_log_level=cfg['General']['log_level'])
+    logging_config(file_log_level=cfg['General']['log_level'],
+                   stream_log_level=log_level)
 
     # http://doc.qt.io/qt-5/highdpi.html
     # https://vicrucann.github.io/tutorials/osg-qt-high-dpi/